// src/shared/api/generated/models/speechVoicePageResponseDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Lola Backend Platform
 * CMS, integration, chat and realtime API for Lola AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { SpeechVoiceResponseDto } from "./speechVoiceResponseDto";

export interface SpeechVoicePageResponseDto {
  hasMore: boolean;
  items: SpeechVoiceResponseDto[];
  /** @nullable */
  nextCursor?: string | null;
}